Current Version
ai can target the player, but the projectiles will not change trajectory
if the player is on higher or lower elevation. level 3 sprites are not
fully implemented yet. To change from level to level, move to the end
and touch the star (might need to move the player start to check because
currently the levels are hard to transverse). To continue to the next
level from the upgrade screen, click the character